Removed the FunctionalAcknowledgment997 Module. Removed and re-added the EDIGlobal module.
Was having compile problems referring to a duplicate module already 
existing for FunctionalAcknowledgment997 Module, so removed the module.

Next compile gave me the same exception for the EDIGlobal Module, so 
copied the source out of the project, deleted EDIGlobal from the project
and re-added EDIGlobal.

Project now successfully compiles.

Needed to remove FunctionalAcknowledgment997 Module, anyway, because we
are going to approach EDI processing in a more generic manner than we
were attempting before.